Move CLI components into submodule
