+++ copy/Cargo.toml
@@ -6,10 +6,8 @@ edition = "2018"
 license = "GPL-3.0-or-later"
 
 [dependencies]
-# FIXME: Move to radicle-bins
-anyhow = "1.0"
+anyhow = "1"
 argh = "0.1"
-
 either = "1.3"
 librad = { path = "../librad" }
 radicle-git-ext = { path = "../git-ext" }
